[refactor] int를 long 수정

+++ back/src/main/java/board/pjt/back/dao/BoardLikeDao.java
@@ -41,7 +41,7 @@ public class BoardLikeDao {
     }
 
     public int toggleBoardLike(UserDetails userDetails, BoardToggleLikeRequestDto requestDto) {
-        int board_id = requestDto.getBoard_id();
+        long board_id = requestDto.getBoard_id();
         String email = userDetails.getUsername();
         requestDto.setCreated_by(email);
         BoardLikeListRequestDto boardLikeListRequestDto = new BoardLikeListRequestDto();
