+++ back/src/main/java/board/pjt/back/dto/category/CategorySelectRequestDto.java
@@ -1,20 +1,20 @@
 package board.pjt.back.dto.category;
 
 public class CategorySelectRequestDto {
-    private Integer category_id;
+    private long category_id;
 
     public CategorySelectRequestDto() {
     }
 
-    public CategorySelectRequestDto(Integer category_id) {
+    public CategorySelectRequestDto(long category_id) {
         this.category_id = category_id;
     }
 
-    public Integer getCategory_id() {
+    public long getCategory_id() {
         return category_id;
     }
 
-    public void setCategory_id(Integer category_id) {
+    public void setCategory_id(long category_id) {
         this.category_id = category_id;
     }
 }
