+++ back/src/main/java/board/pjt/back/dto/thumbnail/DeleteThumbnailRequestDto.java
@@ -3,12 +3,12 @@ package board.pjt.back.dto.thumbnail;
 import java.util.Objects;
 
 public class DeleteThumbnailRequestDto {
-    private int board_id;
+    private long board_id;
 
     public DeleteThumbnailRequestDto() {
     }
 
-    public DeleteThumbnailRequestDto(int board_id) {
+    public DeleteThumbnailRequestDto(long board_id) {
         this.board_id = board_id;
     }
 
@@ -32,11 +32,11 @@ public class DeleteThumbnailRequestDto {
                 '}';
     }
 
-    public int getBoard_id() {
+    public long getBoard_id() {
         return board_id;
     }
 
-    public void setBoard_id(int board_id) {
+    public void setBoard_id(long board_id) {
         this.board_id = board_id;
     }
 }
